Added french settings page and moved button.
