+++ src/user_interface/english_start.ui
@@ -329,44 +329,6 @@ color:rgb(255, 255, 255);</string>
      </property>
     </item>
    </widget>
-   <widget class="QPushButton" name="uploadButton">
-    <property name="geometry">
-     <rect>
-      <x>1049</x>
-      <y>120</y>
-      <width>111</width>
-      <height>41</height>
-     </rect>
-    </property>
-    <property name="styleSheet">
-     <string notr="true">background-color: rgb(0, 85, 127);
-            color: rgb(255, 255, 255);
-            font: 14pt &quot;MS Shell Dlg 2&quot;;
-            border-radius: 10px;</string>
-    </property>
-    <property name="text">
-     <string>Upload</string>
-    </property>
-   </widget>
-   <widget class="QPushButton" name="updateCRKN">
-    <property name="geometry">
-     <rect>
-      <x>875</x>
-      <y>120</y>
-      <width>150</width>
-      <height>41</height>
-     </rect>
-    </property>
-    <property name="styleSheet">
-     <string notr="true">background-color: rgb(0, 85, 127);
-            color: rgb(255, 255, 255);
-            font: 14pt &quot;MS Shell Dlg 2&quot;;
-            border-radius: 10px;</string>
-    </property>
-    <property name="text">
-     <string>Update CRKN</string>
-    </property>
-   </widget>
    <widget class="QLabel" name="universityName">
     <property name="geometry">
      <rect>
